Gen4 unown criteria & Seed generation (#39)
* gen 4 unown form
moves the handling of form requests from ALM into PKHeX criteria so that the unown comes out of the database with the correct form.

* generate from seed
This utilizes PKHeX new GenerateSeed32 and GenerateSeed64 to be able to generate pokemon with a given seed. This is useful for difficult to solve sets or when the user knows their origin seed and wants to generate from that.

To utilize this feature include
~.Seed= in your showdown set
example:
Snover
~.Seed=88A54678f3409003